Add files via upload
Billers at my organization have to rename thousands of PDFs to the customer's name before they can be sent. It costs so much time Analysts and others often have to help. I wrote this script in my free time to eliminate the need to do so manually. 

It does so by breaking the text of each pdf into a data frame, finding the relevant information, and formatting the data to suit a file name. I have added functionality to allow the user to pick how they would like to have the pdfs renamed. Over time I plan to add additionally functionality to this, as time allows.