Create Project Number to Excel
A quick way to grab the project numbers and invoice totals and export them to excel. I built this as a faster way to make sure all invoices matched up with what was expected.

It may be used in the future for verifying project billing versus internal company data on a larger scale.